+++ app/src/main/java/com/example/coffeeapp/domain/main/menu/MenuRepository.kt
@@ -3,5 +3,5 @@ package com.example.coffeeapp.domain.main.menu
 import com.example.coffeeapp.data.main.menu.ShopMenu
 
 interface MenuRepository {
-    suspend operator fun invoke(id: Int?):ArrayList<ShopMenu>
+    suspend operator fun invoke(id: Long?):ArrayList<ShopMenu>
 }
